LK-43, LK-41

+++ src/main/java/ee/ut/math/tvt/salessystem/dataobjects/Purchase.java
@@ -3,23 +3,23 @@ package ee.ut.math.tvt.salessystem.dataobjects;
 import java.math.BigDecimal;
 
 public class Purchase {
-    private int productId;
+    private long productId;
     private String productName;
-    private BigDecimal productPrice;
+    private double productPrice;
     private int quantity;
 
-    public Purchase(int productId, String productName, BigDecimal productPrice, int quantity) {
+    public Purchase(long productId, String productName, double productPrice, int quantity) {
         this.productId = productId;
         this.productName = productName;
         this.productPrice = productPrice;
         this.quantity = quantity;
     }
 
-    public int getProductId() {
+    public long getProductId() {
         return productId;
     }
 
-    public void setProductId(int productId) {
+    public void setProductId(long productId) {
         this.productId = productId;
     }
 
@@ -31,11 +31,11 @@ public class Purchase {
         this.productName = productName;
     }
 
-    public BigDecimal getProductPrice() {
+    public double getProductPrice() {
         return productPrice;
     }
 
-    public void setProductPrice(BigDecimal productPrice) {
+    public void setProductPrice(double productPrice) {
         this.productPrice = productPrice;
     }
